requirements.txt
aiohttp
beautifulsoup4
google-genai
milvus-lite
pymilvus
pymilvus-model
pypdf
python-dotenv
requests
tiktoken
tomli
torch
streamlit
atlassian-python-api